+++ src/main/java/seedu/address/commons/util/StringUtil.java
@@ -15,13 +15,11 @@ public class StringUtil {
         return strings.stream().filter(s -> s.equals(query.toLowerCase())).count() > 0;
     }
     
-    //@@author A0135767U
     public static boolean containsSubstringIgnoreCase(String source, String query) {
         String[] split = source.toLowerCase().split("\\s+");
         List<String> strings = Arrays.asList(split);
         return strings.stream().filter(s -> s.contains(query.toLowerCase())).count() > 0;
     }
-    //@@author
 
     /**
      * Returns a detailed message of the t, including the stack trace.
