+++ src/main/java/seedu/address/logic/commands/ListCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.StringUtil;
 
-//@@author A0135767U
+
 /**
  * Lists all tasks in the SmartyDo to the user.
  */
@@ -19,7 +19,6 @@ public class ListCommand extends Command {
 	private final String ARGUMENT_OVERDUE = "overdue";
 	private final String ARGUMENT_UPCOMING = "upcoming";
 	private final String ARGUMENT_COMPLETED = "completed";
-	private final String ARGUMENT_INCOMPLETE = "incomplete";
 
     public static final String COMMAND_WORD = "list";
 
@@ -49,9 +48,6 @@ public class ListCommand extends Command {
     		break;
     	case ARGUMENT_COMPLETED:
     		model.updateFilteredListToShowCompleted(true);
-    		break;
-    	case ARGUMENT_INCOMPLETE:
-    		model.updateFilteredListToShowCompleted(false);
     		break;
     	default:
         	indicateAttemptToExecuteIncorrectCommand();
