+++ src/main/java/seedu/address/logic/commands/ViewCommand.java
@@ -6,7 +6,6 @@ import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.commons.events.ui.ViewItemRequestEvent;
 import seedu.address.model.task.ReadOnlyTask;
 
-//@@author A0135767U
 /**
  * Lists all tasks in the SmartyDo to the user.
  */
