+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -379,7 +379,6 @@ public class Parser {
         return new FindCommand(keywordSet);
     }
 
-    //@@author A0135767U
     private Command prepareList(String args) {
         final Matcher matcher = KEYWORDS_ARGS_FORMAT.matcher(args.trim());
         if (!matcher.matches()) {
@@ -392,7 +391,6 @@ public class Parser {
         final Set<String> keywordSet = new HashSet<>(Arrays.asList(keywords));
         return new ListCommand(keywordSet);
     }
-    //@@author
 
     /**
      * Parses arguments in the context of the add task command.
