+++ src/main/java/seedu/address/model/Model.java
@@ -30,7 +30,6 @@ public interface Model {
     /** Returns the filtered task list as an {@code UnmodifiableObservableList<ReadOnlyTask>} */
     UnmodifiableObservableList<ReadOnlyTask> getFilteredTaskList();
     
-    //@@author A0135767U
     /** Updates the filter of the filtered task list to show all tasks */
     void updateFilteredListToShowAll();
