+++ src/main/java/seedu/address/ui/UiManager.java
@@ -6,7 +6,6 @@ import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
 import seedu.address.MainApp;
 import seedu.address.commons.core.ComponentManager;
 import seedu.address.commons.core.Config;
@@ -22,7 +21,6 @@ import seedu.address.model.UserPrefs;
 
 import java.util.logging.Logger;
 
-//@@author A0135767U
 /**
  * The manager of the UI component.
  */
@@ -34,7 +32,6 @@ public class UiManager extends ComponentManager implements Ui {
     private Config config;
     private UserPrefs prefs;
     private MainWindow mainWindow;
-    private TaskWindow taskWindow;
 
     public UiManager(Logic logic, Config config, UserPrefs prefs) {
         super();
@@ -50,20 +47,11 @@ public class UiManager extends ComponentManager implements Ui {
 
         //Set the application icon.
         primaryStage.getIcons().add(getImage(ICON_APPLICATION));
-        
-        try {
-            //Set stage to Transparent
-            primaryStage.initStyle(StageStyle.TRANSPARENT);
-        } catch (IllegalStateException e) {
-        	logger.warning(e.getMessage());
-        }
 
         try {
             mainWindow = MainWindow.load(primaryStage, config, prefs, logic);
             mainWindow.show(); //This should be called before creating other UI parts
             mainWindow.fillInnerParts();
-            
-            taskWindow = TaskWindow.load(primaryStage, prefs);
 
         } catch (Throwable e) {
             logger.severe(StringUtil.getDetails(e));
@@ -133,16 +121,14 @@ public class UiManager extends ComponentManager implements Ui {
     @Subscribe
     private void handleTaskPanelSelectionChangedEvent(TaskPanelSelectionChangedEvent event){
         logger.info(LogsCenter.getEventHandlingLogMessage(event));
-        taskWindow.loadTaskPage(event.getNewSelection());
-        taskWindow.show();
+        mainWindow.loadTaskPage(event.getNewSelection());
     }
     
     @Subscribe
     private void handleViewItemRequestEvent(ViewItemRequestEvent event) {
     	logger.info(LogsCenter.getEventHandlingLogMessage(event));
     	mainWindow.getTaskListPanel().scrollDeselect(event.targetIndex);
-    	taskWindow.loadTaskCard(event.getNewSelection());
-    	taskWindow.show();
+    	mainWindow.loadTaskCard(event.getNewSelection());
     }
 
 }
