--- src/main/resources/view/TaskWindow.fxml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import java.net.URL?>
-
-<!-- @@author A0135767U -->
-<VBox maxHeight="Infinity" maxWidth="Infinity" minHeight="-Infinity" minWidth="-Infinity" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="seedu.address.ui.TaskWindow">
-    <stylesheets>
-        <URL value="@DarkTheme.css" />
-        <URL value="@Extensions.css" />
-    </stylesheets>
-    <children>
-        <MenuBar VBox.vgrow="NEVER">
-            <menus>
-                <Menu mnemonicParsing="false" text="Menu">
-                    <items>
-                        <MenuItem mnemonicParsing="false" onAction="#hide" text="Close Window" />
-                    </items>
-                </Menu>
-            </menus>
-        </MenuBar>
-        
-        <AnchorPane VBox.vgrow="NEVER" fx:id="browserPlaceholder" styleClass="anchor-pane-with-border">
-            <padding>
-                <Insets top="5.0" bottom="5.0" left="10.0" right="10.0"/>
-            </padding>
-        </AnchorPane>
-        
-    </children>
-</VBox>
